added state and an rpc listener
